新增部分blservice的桩和驱动，更改部分PO和VO，添加dataservice的桩和驱动

+++ MSOHCode/Client/src/dataservice/hotel_dataservice/Hotel_DataService.java
@@ -31,7 +31,7 @@ public interface Hotel_DataService {
 	
 	public ResultMessage deleteRoomType(RoomPO po);
 	
-	public int getRoomPrice(RoomPO po);
+	public double getRoomPrice(RoomPO po);
 	
 	public String getRoomType(RoomPO po);
